+++ CHANGELOG.md
@@ -1,2 +1,9 @@
+## CurseOfEmbrion [1.1.0]
+- Removed the custom speaker audio files announcing the curse
+- Removed the config variable "enableCustomAudio" as the feature is removed
+- Added a feature where the curse will be shown in red text on the moon's description when landing if it's present
+- Added the config variable "onlyBirdsOnCurse" which determines if only Old Birds can spawn outside when the curse is present
+- Ensure that curse generation is only done host side
+
 ## CurseOfEmbrion [1.0.0]
 - Initial release